+++ packages/kit/test/apps/options/source/pages/slash/child/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import { page } from '$app/stores';
+	import { page } from '$app/state';
 </script>
 
-<h2>{$page.url.pathname}</h2>
+<h2>{page.url.pathname}</h2>