+++ packages/package/test/fixtures/typescript-esnext/expected/Test2.svelte
@@ -1,3 +1,3 @@
-<script>
+<script lang="ts">
 	export let foo;
 </script>